+++ sdk/go/kubernetes/config/config.go
@@ -18,6 +18,16 @@ func GetContext(ctx *pulumi.Context) string {
 	return config.Get(ctx, "kubernetes:context")
 }
 
+// BETA FEATURE - If present and set to true, allow configMaps to be mutated.
+// This feature is in developer preview, and is disabled by default.
+//
+// This config can be specified in the following ways using this precedence:
+// 1. This `enableConfigMapMutable` parameter.
+// 2. The `PULUMI_K8S_ENABLE_CONFIGMAP_MUTABLE` environment variable.
+func GetEnableConfigMapMutable(ctx *pulumi.Context) bool {
+	return config.GetBool(ctx, "kubernetes:enableConfigMapMutable")
+}
+
 // BETA FEATURE - If present and set to true, enable server-side diff calculations.
 // This feature is in developer preview, and is disabled by default.
 //